extract reusable function distributed_train_and_eval
